chore: improve scripts

+++ scripts/git-cleanup.sh
@@ -20,6 +20,11 @@
 
 set -eu
 
+if [ "$(git rev-parse --is-inside-work-tree 2>/dev/null)" != "true" ]; then
+  echo "'${PWD}' is not a git repository" >&2
+  exit 1
+fi
+
 git clean -fdx \
   -e .fleet \
   -e .idea \
